+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 10.2.0 (2024-09-11)
+
+### Improvements
+* CONRAT-29134: Fix the failing tests of ReferralAppTrackingIntegrationSpec
+* CONRAT-34200: Optional App Identifier (aid).
+
 ## 10.2.0 (2024-06-25)
 
 ### Improvements